# cmd/api/docs/swagger.yaml
definitions:
  domain.Admin:
    properties:
      admin_name:
        maxLength: 12
        minLength: 4
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        maxLength: 16
        minLength: 8
        type: string
    required:
    - password
    type: object
  request.Addcount:
    properties:
      count:
        minimum: 1
        type: integer
      product_id:
        type: integer
    required:
    - product_id
    type: object
...
  request.AdminLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.BlockUser:
    properties:
      reason:
        type: string
      user_id:
        type: integer
    type: object
  request.Cartreq:
    properties:
      productId:
        type: integer
    type: object
...
  respondse.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/block:
    patch:
      consumes:
      - application/json
      description: Admin can block a  user
      operationId: block-user
      parameters:
      - description: inputs
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.BlockUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respondse.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can block a user
      tags:
      - Admin
...
  /admin/findall:
    get:
      consumes:
      - application/json
      operationId: FindAllUsers
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: FindAllUsers
      tags:
      - Admin
  /admin/finduser/{user_id}:
    get:
      consumes:
      - application/json
      description: Admin can fetch a specific user details using user id
      operationId: find-user-by-id
      parameters:
      - description: ID of the user to be fetched
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can fetch a specific  user details using id
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: Login admin with details
      operationId: LogInAdmin
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: LoginAdmin
      tags:
      - Admin
  /admin/logout:
    post:
      description: Logout as a user exit from the ecommerce site
      operationId: AdminLogout
      responses:
        "200":
          description: Success
        "400":
          description: Failed
      summary: AdminLogout
      tags:
      - Admin
...
  /admin/signup:
    post:
      consumes:
      - application/json
      description: Save admin with details
      operationId: SaveAdmin
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/domain.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: SaveAdmin
      tags:
      - Admin
  /admin/unblock/{user_id}:
    patch:
      consumes:
      - application/json
      description: Admin can unblock a blocked user
      operationId: unblock-user
      parameters:
      - description: ID of the user to be unblocked
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can unblock a blocked user
      tags:
      - Admin
  /cart/Addcount:
    put:
      consumes:
      - application/json
      description: user can delete their cartitems by id
      operationId: Add-Qantity
      parameters:
      - description: Input Field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.Addcount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: User can delete a item
      tags:
      - UsersCart
  /cart/add/item:
    post:
      consumes:
      - application/json
      description: user can add stock of product to user cart
      parameters:
      - description: Input Field info
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.Cartreq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: api for adding product to user cart
      tags:
      - UsersCart
  /cart/remove/item:
    delete:
      consumes:
      - application/json
      description: user can reduce stock of product to user cart
      parameters:
      - description: Input Field info
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/request.Cartreq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: api for removing product from user cart
      tags:
      - UsersCart
  /cart/viewcart:
    get:
      description: user can see all productItem that stored in cart
      operationId: Cart
      responses:
        "200":
          description: successfully got user cart items
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: faild to get cart items
          schema:
            $ref: '#/definitions/respondse.Response'
      security:
      - ApiKeyAuth: []
      summary: api for get all cart item of user
      tags:
      - UsersCart
...
swagger: "2.0"